pull toggle button into its own file
